chore: add error message in feature files can't be found

+++ startTests.js
@@ -1,4 +1,5 @@
 const shell = require("shelljs");
+const fs = require("fs");
 
 // Copy the feature files to where maven can recognise them
 // This requires the openapi-forge to have been installed locally:
@@ -7,16 +8,27 @@ const shell = require("shelljs");
 // | - features/
 // | openapi-forge-java
 // | - features/
-shell.exec("cp ../openapi-forge/features/*.feature features/src/test/resources/com/openapi/forge");
+if (fs.existsSync("../openapi-forge/features/")) {
+  shell.exec(
+    "cp ../openapi-forge/features/*.feature features/src/test/resources/com/openapi/forge"
+  );
 
-// Go into the test folder
-shell.cd("features/");
+  // Go into the test folder
+  shell.cd("features/");
 
-// Pass over control to maven test
-shell.exec("sh ./mvnw test");
+  // Pass over control to maven test
+  shell.exec("sh ./mvnw test");
 
-// Return to original location
-shell.cd("..");
+  // Return to original location
+  shell.cd("..");
 
-// Clean the resources folder
-shell.exec("rm features/src/test/resources/com/openapi/forge/*.feature")
+  // Clean the resources folder
+  shell.exec("rm features/src/test/resources/com/openapi/forge/*.feature");
+} else {
+  shell.echo(
+    "Expecting .feature files to be found at ../openapi-forge/features/ (relative from here)."
+  );
+  shell.echo(
+    "Have you installed the openapi forge project in the correct location?"
+  );
+}
